// src/types.ts
export interface Message {
  _data: {
    id: {
      fromMe: boolean;
      remote: string;
      id: string;
      _serialized: string;
    };
    viewed: boolean;
    body: string;
    type: string;
    t: number;
    notifyName: string;
    from: string;
    to: string;
    ack: number;
    invis: boolean;
    isNewMsg: boolean;
    star: boolean;
    kicNotified: boolean;
    recvFresh: boolean;
    isFromTemplate: boolean;
    thumbnail: string;
    richPreviewType: number;
    pollInvalidated: boolean;
    isSentCagPollCreation: boolean;
    latestEditMsgKey: string | null;
    latestEditSenderTimestampMs: number | null;
    mentionedJidList: string[];
    groupMentions: string[];
    isEventCanceled: boolean;
    eventInvalidated: boolean;
    isVcardOverMmsDocument: boolean;
    isForwarded: boolean;
    labels: string[];
    hasReaction: boolean;
    inviteGrpType: string;
    productHeaderImageRejected: boolean;
    lastPlaybackProgress: number;
    isDynamicReplyButtonsMsg: boolean;
    isCarouselCard: boolean;
    parentMsgId: string | null;
    isMdHistoryMsg: boolean;
    stickerSentTs: number;
    isAvatar: boolean;
    lastUpdateFromServerTs: number;
    invokedBotWid: string | null;
    bizBotType: string | null;
    botResponseTargetId: string | null;
    botPluginType: string | null;
    botPluginReferenceIndex: number | null;
    botPluginSearchProvider: string | null;
    botPluginSearchUrl: string | null;
    botPluginSearchQuery: string | null;
    botPluginMaybeParent: boolean;
    botReelPluginThumbnailCdnUrl: string | null;
    botMsgBodyType: string | null;
    requiresDirectConnection: boolean;
    bizContentPlaceholderType: string | null;
    hostedBizEncStateMismatch: boolean;
    senderOrRecipientAccountTypeHosted: boolean;
    placeholderCreatedWhenAccountIsHosted: boolean;
    links: string[];
  };
  mediaKey?: string;
  id: {
    fromMe: boolean;
    remote: string;
    id: string;
    _serialized: string;
  };
  ack: number;
  hasMedia: boolean;
  body: string;
  type: string;
  timestamp: number;
  from: string;
  to: string;
  author?: string;
  deviceType: string;
  isForwarded: boolean;
  forwardingScore: number;
  isStatus: boolean;
  isStarred: boolean;
  broadcast?: boolean;
  fromMe: boolean;
  hasQuotedMsg: boolean;
  hasReaction: boolean;
  duration?: number;
  location?: string;
  vCards: string[];
  inviteV4?: string;
  mentionedIds: string[];
  groupMentions: string[];
  orderId?: string;
  token?: string;
  isGif: boolean;
  isEphemeral?: boolean;
  links: string[];
}